+++ scout/server/blueprints/dashboard/views.py
@@ -1,58 +1,47 @@
+import logging
+
+from pprint import pprint as pp
+
 from flask import (abort, Blueprint, current_app, redirect, render_template,
                    request, url_for, send_from_directory, jsonify, flash)
 from flask_login import current_user
 
 from scout.server.extensions import store
 
+from .controllers import get_dashboard_info
+
 blueprint = Blueprint('dashboard', __name__, template_folder='templates')
 
+LOG = logging.getLogger(__name__)
 
-@blueprint.route('/dashboard')
+@blueprint.route('/dashboard', methods=['GET', 'POST'])
 def index():
     """Display the Scout dashboard."""
-    total_cases = store.cases().count()
+    accessible_institutes = None
+    if not 'admin' in current_user.roles:
+        accessible_institutes = current_user.institutes
+        if not accessible_institutes:
+            flash('Not allowed to see information - please visit the dashboard later!')
+            return redirect(url_for('cases.index'))
+
+    institutes = [inst for inst in store.institutes(accessible_institutes)]
+    # Insert a entry that displays all institutes in the beginning of the array
+    institutes.insert(0, {'_id': None, 'display_name': 'All institutes'})
+
+    institute_id = None
+    if request.method=='POST':
+        institute_id = request.form.get('institute')
+
+    LOG.info("Fetch all cases with institute: %s", institute_id)
+    data = get_dashboard_info(store, institute_id)
+    data['institutes'] = institutes
+    data['choice'] = institute_id
+    total_cases = data['total_cases']
+    
+    LOG.info("Found %s cases", total_cases)
     if total_cases == 0:
-        flash('no cases loaded - please visit the dashboard later!', 'info')
+        flash('no cases loaded for institute {} - please visit the dashboard later!'.format(institute_id), 'info')
         return redirect(url_for('cases.index'))
-    cases = [{'status': 'all', 'count': total_cases, 'percent': 1}]
-    query = store.case_collection.aggregate([
-        {'$group' : {'_id': '$status', 'count': {'$sum': 1}}}
-    ])
-    for status_group in query:
-        cases.append({'status': status_group['_id'],
-                      'count': status_group['count'],
-                      'percent': status_group['count'] / total_cases})
-
-    query = store.case_collection.aggregate([
-        {'$unwind': '$individuals'},
-        {'$group': {'_id': '$individuals.analysis_type', 'count': {'$sum': 1}}}
-    ])
-    analysis_types = [{'name': group['_id'], 'count': group['count']} for group in query]
-
-    phenotype_terms = store.case_collection.find({'phenotype_terms.1': {'$exists': True}}).count()
-    causative_variants = store.case_collection.find({'causatives.1': {'$exists': True}}).count()
-    pinned_variants = store.case_collection.find({'suspects.1': {'$exists': True}}).count()
-    cohort_tags = store.case_collection.find({'cohorts.1': {'$exists': True}}).count()
-    overview = [{
-        'title': 'Phenotype terms',
-        'count': phenotype_terms,
-        'percent': phenotype_terms / total_cases,
-    }, {
-        'title': 'Causative variants',
-        'count': causative_variants,
-        'percent': causative_variants / total_cases,
-    }, {
-        'title': 'Pinned variants',
-        'count': pinned_variants,
-        'percent': pinned_variants / total_cases,
-    }, {
-        'title': 'Cohort tag',
-        'count': cohort_tags,
-        'percent': cohort_tags / total_cases,
-    }]
+    
     return render_template(
-        'dashboard/index.html',
-        cases=cases,
-        analysis_types=analysis_types,
-        overview=overview,
-    )
+        'dashboard/index.html', **data)